refactor: api 관련 상수 파일 분리 - constants/api 폴대 생성 및 내부에 관련 상수 이동 및 적용
