+++ frontend/src/api/channels.ts
@@ -1,6 +1,6 @@
 import { fetcher } from "@src/api";
 
-import { API_ENDPOINT } from "@src/@constants";
+import { API_ENDPOINT } from "@src/@constants/api";
 import { ResponseChannels, ResponseSubscribedChannels } from "@src/@types/api";
 
 import { getPrivateHeaders } from "@src/api/utils";
