+++ frontend/src/api/messages.ts
@@ -1,6 +1,6 @@
 import { fetcher } from "@src/api";
 
-import { API_ENDPOINT } from "@src/@constants";
+import { API_ENDPOINT } from "@src/@constants/api";
 import { ResponseMessages } from "@src/@types/api";
 
 import { getPrivateHeaders } from "@src/api/utils";
