+++ frontend/src/components/Router/PrivateRouter/index.tsx
@@ -5,7 +5,8 @@ import { Outlet } from "react-router-dom";
 
 import Loader from "@src/components/Loader";
 
-import { PATH_NAME, QUERY_KEY } from "@src/@constants";
+import { PATH_NAME } from "@src/@constants";
+import { QUERY_KEY } from "@src/@constants/api";
 
 import { isCertificated } from "@src/api/auth";
