+++ frontend/src/hooks/@query/useGetSubscribedChannels.ts
@@ -1,6 +1,6 @@
 import { UseQueryResult, useQuery } from "react-query";
 
-import { QUERY_KEY } from "@src/@constants";
+import { QUERY_KEY } from "@src/@constants/api";
 import { ResponseSubscribedChannels } from "@src/@types/api";
 import { CustomError } from "@src/@types/shared";
