+++ frontend/src/hooks/useAuthentication.ts
@@ -4,12 +4,8 @@ import { useNavigate } from "react-router-dom";
 import useRecentFeedPath from "@src/hooks/useRecentFeedPath";
 import useSnackbar from "@src/hooks/useSnackbar";
 
-import {
-  ACCESS_TOKEN_KEY,
-  MESSAGES,
-  PATH_NAME,
-  QUERY_KEY,
-} from "@src/@constants";
+import { MESSAGES, PATH_NAME } from "@src/@constants";
+import { ACCESS_TOKEN_KEY, QUERY_KEY } from "@src/@constants/api";
 import { deleteCookie, setCookie } from "@src/@utils";
 
 interface UseAuthenticationResult {
